refactor: remove `errors.go` file
